pyproject.toml
[project]
name = "github-security-scanner"
version = "1.0.0"
description = "Ferramenta de análise de segurança automatizada para organizações GitHub"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Security Team", email = "security@example.com"}
]
keywords = ["security", "github", "scanner", "secrets", "vulnerabilities", "sast"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "httpx>=0.25",
    "pydantic[email]>=2.0",
    "pydantic-settings>=2.0",
    "email-validator>=2.0",
    "gitpython>=3.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "aiofiles>=23.0",
    "tenacity>=8.2",
    "python-dotenv>=1.0",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    # Database
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",
    # Redis
    "redis>=5.0",
    # Auth
    "pyjwt>=2.8",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.1",
    "passlib[bcrypt]>=1.7",
    # Async utilities
    "aiocache>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "types-pyyaml",
    "types-aiofiles",
]

[project.scripts]
github-security-scanner = "github_security_scanner.__main__:main"
gss = "github_security_scanner.__main__:main"
gss-api = "github_security_scanner.api:main"

[project.urls]
Homepage = "https://github.com/your-org/github-security-scanner"
Documentation = "https://github.com/your-org/github-security-scanner/docs"
Repository = "https://github.com/your-org/github-security-scanner"
Issues = "https://github.com/your-org/github-security-scanner/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_security_scanner"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
]
ignore = [
    "S101",   # assert usage
    "S105",   # hardcoded password (we detect these, not avoid them)
    "S106",   # hardcoded password
    "S107",   # hardcoded password
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/github_security_scanner --cov-report=term-missing"

[tool.coverage.run]
source = ["src/github_security_scanner"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

